+++ enterprise/internal/campaigns/syncer/util.go
@@ -1,4 +1,4 @@
-package campaigns
+package syncer
 
 import (
 	"context"
